LPD-35022 Extracting code not related to the staging page per se
